+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Use config struct to initialize `Tiingo` reader ([#3](https://github.com/AleksanderWWW/go-datareader/pull/3))
 - Remove logging to file ([#6](https://github.com/AleksanderWWW/go-datareader/pull/6))
 - Move base urls to a separate `constants.go` file ([#8](https://github.com/AleksanderWWW/go-datareader/pull/8))
+- Move configs to a separate file ([#10](https://github.com/AleksanderWWW/go-datareader/pull/10))
 
 ### Fixes
 - Handle empty dataframe list in `concatDataframes` ([#9](https://github.com/AleksanderWWW/go-datareader/pull/9))